+++ Demos/04-Teams/Demo-06/MessagingExtToolkit/MySearchExt/.env
@@ -1,0 +1,2 @@
+BotId=67d74167-4541-42ad-b19b-7e1a9eacd37a
+BotPassword=UI.~db3~vwi0223Wyl8b.BsX0~0X5ASU89
