+++ Demos/04-Teams/Demo-06/MessagingExtToolkit/MySearchExt/appPackage/color.png
