+++ Demos/04-Teams/Demo-06/MessagingExtToolkit/MySearchExt/appPackage/manifest.json
@@ -2,7 +2,7 @@
     "$schema": "https://developer.microsoft.com/en-us/json-schemas/teams/v1.8/MicrosoftTeams.schema.json",
     "manifestVersion": "1.8",
     "version": "1.0.0",
-    "id": "672bb6d5-4786-48f8-8c2d-03490dc6cfc8",
+    "id": "811a087c-a704-40f0-a859-ffbded7c327c",
     "packageName": "com.microsoft.teams.extension",
     "developer": {
         "name": "Teams App, Inc.",
@@ -15,7 +15,7 @@
         "outline": "outline.png"
     },
     "name": {
-        "short": "SearchBasedMsgExt",
+        "short": "MySearchExt",
         "full": ""
     },
     "description": {
@@ -25,7 +25,7 @@
     "accentColor": "#FFFFFF",
     "composeExtensions": [
         {
-            "botId": "975dfaca-6c46-4c25-9308-989933de114f",
+            "botId": "67d74167-4541-42ad-b19b-7e1a9eacd37a",
             "canUpdateConfiguration": true,
             "commands": [
                 {
