--- Demos/04-Teams/Demo-06/MessagingExtToolkit/SearchBasedMsgExt/.env
@@ -1,2 +1,0 @@
-BotId=975dfaca-6c46-4c25-9308-989933de114f
-BotPassword=G-uIM8qD43YnDOtAeh-10h4_--9K~V~sI1
